requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python
redis
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
